BuyerService/requirements.txt
anyio==3.6.1
autopep8==1.7.0
click==8.1.3
dependency-injector==4.40.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.18.3
